+++ Project.toml
@@ -15,6 +15,7 @@ IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
 IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
 Lerche = "d42ef402-04e6-4356-9f73-091573ea58dc"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 NLsolve = "2774e3e8-f4cf-5e23-947b-6d7e65073b56"
 Optim = "429524aa-4258-5aef-a3af-852621145aeb"
