+++ examples/models/rbc.yaml
@@ -79,7 +79,7 @@ exogenous:
 
 options:
     grid: !Cartesian
-        orders: [1000]
+        orders: [20]
 
 test: |
     chi*n^eta*c^sigma - (1-alpha)*y/n           | 0 <= n <= inf
