+++ src/Dolo.jl
@@ -153,9 +153,10 @@ include("numeric/decision_rules/smolyak.jl")
 include("numeric/decision_rules/complete.jl")
 
 include("algos/steady_state.jl")
-include("algos/time_iteration.jl")
 include("algos/improved_time_iteration.jl")
+include("algos/time_iteration.jl")
 include("algos/time_iteration_direct.jl")
+include("algos/time_iteration_old.jl")
 include("algos/value_iteration.jl")
 include("algos/perturbation.jl")
 include("algos/simulation.jl")
